Revert "adding intralist diversity metric (#15)"
This reverts commit 2238b23d5563500eb18ef649dc46a5c70a064adc.
