+++ docs/.buildinfo
@@ -1,4 +1,4 @@
 # Sphinx build info version 1
 # This file hashes the configuration used when building these files. When it is not found, a full rebuild will be done.
-config: 81c8dda9601647e27bd83929d7a49f75
+config: a4efaf5788f180055a98697641fd0ae0
 tags: 645f666f9bcd5a90fca523b33c5a78b7
