+++ docs/about_reco.html
@@ -4,7 +4,7 @@
   <meta charset="utf-8" /><meta name="generator" content="Docutils 0.17.1: http://docutils.sourceforge.net/" />
 
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <title>About Recommenders Metrics &mdash; Jurity 1.3.2 documentation</title>
+  <title>About Recommenders Metrics &mdash; Jurity 1.3.1 documentation</title>
       <link rel="stylesheet" href="_static/pygments.css" type="text/css" />
       <link rel="stylesheet" href="_static/css/theme.css" type="text/css" />
   <!--[if lt IE 9]>
@@ -15,7 +15,7 @@
         <script src="_static/jquery.js"></script>
         <script src="_static/underscore.js"></script>
         <script src="_static/doctools.js"></script>
-        <script async="async" src="https://cdn.jsdelivr.net/npm/mathjax@3/es5/tex-mml-chtml.js"></script>
+        <script defer="defer" src="https://cdn.jsdelivr.net/npm/mathjax@3/es5/tex-mml-chtml.js"></script>
     <script src="_static/js/theme.js"></script>
     <link rel="index" title="Index" href="genindex.html" />
     <link rel="search" title="Search" href="search.html" />
@@ -57,7 +57,6 @@
 </li>
 <li class="toctree-l2"><a class="reference internal" href="#diversity-recommender-metrics">Diversity Recommender Metrics</a><ul>
 <li class="toctree-l3"><a class="reference internal" href="#inter-list-diversity">Inter-List Diversity</a></li>
-<li class="toctree-l3"><a class="reference internal" href="#intra-list-diversity">Intra-List Diversity</a></li>
 </ul>
 </li>
 </ul>
@@ -185,16 +184,6 @@ user set, <span class="math notranslate nohighlight">\(i, j \in \{0,1,\cdots,N\}
 recommendations for <span class="math notranslate nohighlight">\(u_i\)</span>. <span class="math notranslate nohighlight">\(I\)</span> is the set of all unique user pairs, <span class="math notranslate nohighlight">\(\forall~i&lt;j, \{u_i, u_j\} \in I\)</span>.</p>
 <div class="math notranslate nohighlight">
 \[Inter\mbox{-}list~diversity = \frac{\sum_{i,j, \{u_i, u_j\} \in I}(cosine\_distance(R_{u_i}, R_{u_j}))}{|I|}\]</div>
-</section>
-<section id="intra-list-diversity">
-<h3>Intra-List Diversity<a class="headerlink" href="#intra-list-diversity" title="Permalink to this headline"></a></h3>
-<p>Intra-List Diversity&#64;k measures the intra-list diversity of the recommendations when only k recommendations are
-made to the user. It measures how the items in the item list for each user are different from each other. This metric has a
-range in <span class="math notranslate nohighlight">\([0, 1]\)</span>. The higher this metric is, the more diversified items are recommended to
-users. Let <span class="math notranslate nohighlight">\(U\)</span> denote the set of <span class="math notranslate nohighlight">\(N\)</span> unique users, <span class="math notranslate nohighlight">\(u_i\)</span> denotes the i-th user in the user set,
-<span class="math notranslate nohighlight">\(v_p^{u_i}\)</span> are the item features of the p-th recommended item for user <span class="math notranslate nohighlight">\(u_i\)</span> and p &lt; q.</p>
-<div class="math notranslate nohighlight">
-\[Intra \mbox{-} list~diversity = 1 - \frac{1}{U}\sum_{i=1}^U average(consine\_similarity(v_p^{u_i}, v_q^{u_i}))\]</div>
 </section>
 </section>
 </section>
