+++ docs/genindex.html
@@ -3,7 +3,7 @@
 <head>
   <meta charset="utf-8" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <title>Index &mdash; Jurity 1.3.2 documentation</title>
+  <title>Index &mdash; Jurity 1.3.1 documentation</title>
       <link rel="stylesheet" href="_static/pygments.css" type="text/css" />
       <link rel="stylesheet" href="_static/css/theme.css" type="text/css" />
   <!--[if lt IE 9]>
@@ -71,11 +71,13 @@
 <div class="genindex-jumpbox">
  <a href="#B"><strong>B</strong></a>
  | <a href="#D"><strong>D</strong></a>
+ | <a href="#F"><strong>F</strong></a>
  | <a href="#G"><strong>G</strong></a>
  | <a href="#J"><strong>J</strong></a>
  | <a href="#M"><strong>M</strong></a>
  | <a href="#R"><strong>R</strong></a>
  | <a href="#S"><strong>S</strong></a>
+ | <a href="#T"><strong>T</strong></a>
  
 </div>
 <h2 id="B">B</h2>
@@ -97,10 +99,10 @@
 </li>
       <li><a href="api.html#jurity.fairness.BinaryFairnessMetrics.AverageOdds">BinaryFairnessMetrics.AverageOdds (class in jurity.fairness)</a>
 </li>
-  </ul></td>
-  <td style="width: 33%; vertical-align: top;"><ul>
       <li><a href="api.html#jurity.fairness.BinaryFairnessMetrics.EqualOpportunity">BinaryFairnessMetrics.EqualOpportunity (class in jurity.fairness)</a>
 </li>
+  </ul></td>
+  <td style="width: 33%; vertical-align: top;"><ul>
       <li><a href="api.html#jurity.fairness.BinaryFairnessMetrics.FNRDifference">BinaryFairnessMetrics.FNRDifference (class in jurity.fairness)</a>
 </li>
       <li><a href="api.html#jurity.fairness.BinaryFairnessMetrics.GeneralizedEntropyIndex">BinaryFairnessMetrics.GeneralizedEntropyIndex (class in jurity.fairness)</a>
@@ -108,6 +110,10 @@
       <li><a href="api.html#jurity.fairness.BinaryFairnessMetrics.PredictiveEquality">BinaryFairnessMetrics.PredictiveEquality (class in jurity.fairness)</a>
 </li>
       <li><a href="api.html#jurity.fairness.BinaryFairnessMetrics.TheilIndex">BinaryFairnessMetrics.TheilIndex (class in jurity.fairness)</a>
+</li>
+      <li><a href="api.html#jurity.mitigation.BinaryMitigation">BinaryMitigation (class in jurity.mitigation)</a>
+</li>
+      <li><a href="api.html#jurity.mitigation.BinaryMitigation.EqualizedOdds">BinaryMitigation.EqualizedOdds (class in jurity.mitigation)</a>
 </li>
       <li><a href="api.html#jurity.recommenders.BinaryRecoMetrics">BinaryRecoMetrics (class in jurity.recommenders)</a>
 </li>
@@ -133,7 +139,17 @@
 </li>
       <li><a href="api.html#jurity.recommenders.DiversityRecoMetrics.InterListDiversity">DiversityRecoMetrics.InterListDiversity (class in jurity.recommenders)</a>
 </li>
-      <li><a href="api.html#jurity.recommenders.DiversityRecoMetrics.IntraListDiversity">DiversityRecoMetrics.IntraListDiversity (class in jurity.recommenders)</a>
+  </ul></td>
+</tr></table>
+
+<h2 id="F">F</h2>
+<table style="width: 100%" class="indextable genindextable"><tr>
+  <td style="width: 33%; vertical-align: top;"><ul>
+      <li><a href="api.html#jurity.mitigation.BinaryMitigation.EqualizedOdds.fit">fit() (jurity.mitigation.BinaryMitigation.EqualizedOdds method)</a>
+</li>
+  </ul></td>
+  <td style="width: 33%; vertical-align: top;"><ul>
+      <li><a href="api.html#jurity.mitigation.BinaryMitigation.EqualizedOdds.fit_transform">fit_transform() (jurity.mitigation.BinaryMitigation.EqualizedOdds method)</a>
 </li>
   </ul></td>
 </tr></table>
@@ -175,8 +191,6 @@
         <li><a href="api.html#jurity.recommenders.BinaryRecoMetrics.CTR.get_score">(jurity.recommenders.BinaryRecoMetrics.CTR method)</a>
 </li>
         <li><a href="api.html#jurity.recommenders.DiversityRecoMetrics.InterListDiversity.get_score">(jurity.recommenders.DiversityRecoMetrics.InterListDiversity method)</a>
-</li>
-        <li><a href="api.html#jurity.recommenders.DiversityRecoMetrics.IntraListDiversity.get_score">(jurity.recommenders.DiversityRecoMetrics.IntraListDiversity method)</a>
 </li>
         <li><a href="api.html#jurity.recommenders.RankingRecoMetrics.MAP.get_score">(jurity.recommenders.RankingRecoMetrics.MAP method)</a>
 </li>
@@ -247,6 +261,14 @@
         <li><a href="api.html#jurity.fairness.MultiClassFairnessMetrics.StatisticalParity">(jurity.fairness.MultiClassFairnessMetrics attribute)</a>
 </li>
       </ul></li>
+  </ul></td>
+</tr></table>
+
+<h2 id="T">T</h2>
+<table style="width: 100%" class="indextable genindextable"><tr>
+  <td style="width: 33%; vertical-align: top;"><ul>
+      <li><a href="api.html#jurity.mitigation.BinaryMitigation.EqualizedOdds.transform">transform() (jurity.mitigation.BinaryMitigation.EqualizedOdds method)</a>
+</li>
   </ul></td>
 </tr></table>
