+++ docs/searchindex.js
@@ -1,1 +1,1 @@
-Search.setIndex({docnames:["about_fairness","about_reco","api","index","install","quick"],envversion:{"sphinx.domains.c":2,"sphinx.domains.changeset":1,"sphinx.domains.citation":1,"sphinx.domains.cpp":4,"sphinx.domains.index":1,"sphinx.domains.javascript":2,"sphinx.domains.math":2,"sphinx.domains.python":3,"sphinx.domains.rst":2,"sphinx.domains.std":2,sphinx:56},filenames:["about_fairness.rst","about_reco.rst","api.rst","index.rst","install.rst","quick.rst"],objects:{"":{jurity:[2,0,0,"-"]},"jurity.classification":{BinaryClassificationMetrics:[2,1,1,""]},"jurity.classification.BinaryClassificationMetrics":{AUC:[2,1,1,""],Accuracy:[2,1,1,""],F1:[2,1,1,""],Precision:[2,1,1,""],Recall:[2,1,1,""]},"jurity.classification.BinaryClassificationMetrics.AUC":{get_score:[2,2,1,""]},"jurity.classification.BinaryClassificationMetrics.Accuracy":{get_score:[2,2,1,""]},"jurity.classification.BinaryClassificationMetrics.F1":{get_score:[2,2,1,""]},"jurity.classification.BinaryClassificationMetrics.Precision":{get_score:[2,2,1,""]},"jurity.classification.BinaryClassificationMetrics.Recall":{get_score:[2,2,1,""]},"jurity.fairness":{BinaryFairnessMetrics:[2,1,1,""],MultiClassFairnessMetrics:[2,1,1,""]},"jurity.fairness.BinaryFairnessMetrics":{AverageOdds:[2,1,1,""],DisparateImpact:[2,3,1,""],EqualOpportunity:[2,1,1,""],FNRDifference:[2,1,1,""],GeneralizedEntropyIndex:[2,1,1,""],PredictiveEquality:[2,1,1,""],StatisticalParity:[2,3,1,""],TheilIndex:[2,1,1,""],get_all_scores:[2,2,1,""]},"jurity.fairness.BinaryFairnessMetrics.AverageOdds":{get_score:[2,2,1,""]},"jurity.fairness.BinaryFairnessMetrics.EqualOpportunity":{get_score:[2,2,1,""]},"jurity.fairness.BinaryFairnessMetrics.FNRDifference":{get_score:[2,2,1,""]},"jurity.fairness.BinaryFairnessMetrics.GeneralizedEntropyIndex":{get_score:[2,2,1,""]},"jurity.fairness.BinaryFairnessMetrics.PredictiveEquality":{get_score:[2,2,1,""]},"jurity.fairness.BinaryFairnessMetrics.TheilIndex":{get_score:[2,2,1,""]},"jurity.fairness.MultiClassFairnessMetrics":{DisparateImpact:[2,3,1,""],StatisticalParity:[2,3,1,""],get_all_scores:[2,2,1,""]},"jurity.recommenders":{BinaryRecoMetrics:[2,1,1,""],DiversityRecoMetrics:[2,1,1,""],RankingRecoMetrics:[2,1,1,""]},"jurity.recommenders.BinaryRecoMetrics":{AUC:[2,1,1,""],CTR:[2,1,1,""]},"jurity.recommenders.BinaryRecoMetrics.AUC":{get_score:[2,2,1,""]},"jurity.recommenders.BinaryRecoMetrics.CTR":{get_score:[2,2,1,""]},"jurity.recommenders.DiversityRecoMetrics":{InterListDiversity:[2,1,1,""],IntraListDiversity:[2,1,1,""]},"jurity.recommenders.DiversityRecoMetrics.InterListDiversity":{get_score:[2,2,1,""]},"jurity.recommenders.DiversityRecoMetrics.IntraListDiversity":{get_score:[2,2,1,""]},"jurity.recommenders.RankingRecoMetrics":{MAP:[2,1,1,""],NDCG:[2,1,1,""],Precision:[2,1,1,""],Recall:[2,1,1,""]},"jurity.recommenders.RankingRecoMetrics.MAP":{get_score:[2,2,1,""]},"jurity.recommenders.RankingRecoMetrics.NDCG":{get_score:[2,2,1,""]},"jurity.recommenders.RankingRecoMetrics.Precision":{get_score:[2,2,1,""]},"jurity.recommenders.RankingRecoMetrics.Recall":{get_score:[2,2,1,""]}},objnames:{"0":["py","module","Python module"],"1":["py","class","Python class"],"2":["py","method","Python method"],"3":["py","attribute","Python attribute"]},objtypes:{"0":"py:module","1":"py:class","2":"py:method","3":"py:attribute"},terms:{"0":[0,1,2,3,5],"1":[0,1,2,3,5],"10":2,"10000":2,"1103":[1,2],"12":2,"122":2,"16":0,"2":[0,2,3,5],"2011":[1,2],"2016":0,"2018":[0,2],"28th":[1,2],"3":[2,3,4,5],"30":[3,5],"30th":0,"316":2,"4":[2,3,5],"453":2,"4601":[1,2],"5":[1,3,5],"6":[3,4,5],"65":2,"68":2,"7":[3,5],"8":[3,5],"boolean":[1,2],"case":[0,1,2],"class":[2,3,5],"default":2,"dud\u00edk":[1,2],"float":2,"function":[1,2],"import":[1,2,3,5],"int":2,"public":3,"return":[1,2],"static":2,"true":[0,2],"while":[1,2],A:[0,1,2],At:[1,2],By:2,For:1,If:[1,2],In:[0,1,2],It:[1,2],That:2,The:[1,2,3,4],There:[1,2,4],To:4,With:[0,2],__k__:1,_a:1,_basebinaryfair:2,_baserecommend:2,_distanc:[1,2],_i:[0,2],_item_id_column:2,_similar:1,_user_id_column:2,a_i:[1,2],abil:2,abl:1,about:3,abov:2,account:[1,2],accumul:2,accuraci:[0,1,2,3,5],achiev:0,acm:[0,2],across:[0,2],actual:[1,2,3,5],actual_respons:2,actual_responses_batch:2,actual_result:2,addit:2,adher:3,after:[3,4,5],agre:1,algorithm:[2,3],alia:2,all:[1,2,4],allow:2,alpha:[0,2],also:[1,2,3,5],altern:4,alwai:1,among:2,amount:1,an:[1,2],analog:2,ani:[2,4],anyth:2,api:3,appear:1,approach:[0,2],approxim:2,ar:[1,2,4],area:[2,3],artifici:3,arxiv:[1,2],associ:2,assum:2,attribut:[0,2],auc:[2,3,5],auc_acc:2,auc_batch:2,auxiliari:2,avail:[1,2,5],averag:[2,3],averageodd:2,avoid:1,b:[0,2,3,5],b_:[0,2],b_i:[0,2],balanc:2,base:[1,2],batch:2,batch_accumul:2,bdist_wheel:4,befor:[3,5],begin:[0,2],being:[1,2],below:[0,1],best:2,between:[0,1,2],bia:0,binari:[0,3,5],binary_predict:[3,5],binaryclassificationmetr:[2,3,5],binarydisparateimpact:2,binaryfairnessmetr:[2,3,5],binarymitig:[3,5],binaryrecometr:[2,3,5],binarystatisticalpar:2,black:2,bool:2,both:1,bound:[3,5],box:2,build:4,c:[1,3,5],calcul:[0,1,2],call:2,callabl:2,can:[1,2,4],cannot:2,cap:[1,2],captur:[1,2],cdot:1,center:3,characterist:2,chosen:1,classif:0,classifi:[0,2],click:[2,3,5],click_column:[2,3,5],clone:4,close:2,column:2,combin:[1,2],come:1,command:4,commun:1,comput:2,condit:[0,2],confer:[0,1,2],confirm:4,consin:1,consine_similar:2,consist:[1,2],constrain:0,contain:2,context:[1,2],contribut:2,control:2,core:2,correct:[1,2],correctli:2,cosin:[1,2],ctr:[2,3,5],ctr_acc:2,ctr_batch:2,cumul:[2,3],current:2,curv:[2,3],d:[0,2],data:[0,1,2,3,5],datafram:[2,3,5],dataset:[0,2],decis:[0,2],defin:[0,2],definit:[0,1],denot:[0,1,2],depend:[2,4],descript:[3,5],design:2,determin:2,develop:3,dict:2,didn:1,differ:[1,2,3],direct:[1,2],directli:[1,2],discount:[2,3],discov:4,discoveri:[0,2],dispar:3,disparate_impact:2,disparateimpact:2,dist:4,distanc:2,distribut:2,divers:3,diversifi:[1,2],diversityrecometr:2,divid:1,done:1,doubli:[1,2],dr:[1,2],draw:2,due:1,e:[0,2],each:[0,1,2],eas:2,easili:2,edg:1,effect:1,els:2,end:[0,2],enough:1,entropi:[2,3],equal:[2,3],equalizedodd:[3,5],equalopportun:2,equat:2,eric:0,error:1,especi:1,estim:[1,2],evalu:[1,2],even:1,exampl:[0,2,4],excel:3,extend:2,f1:[2,3,5],f1_score:[3,5],f:[1,2],fair_likelihood:[3,5],fair_predict:[3,5],fals:[0,2,3],fd8522f660ee5aeb84394b4f1b329a1d662dc7af:[],featur:[1,2],fidel:3,filter:1,find:2,first:4,flip:0,fn:2,fnr:[2,3],fnrdiffer:2,follow:4,footprint:2,foral:1,formal:2,formula:2,fp:2,fpr:[0,2],fpr_:[0,2],frac:[0,1,2],fraction:2,frame:2,from:[1,2,3,5],full:2,further:1,g:[0,2],gain:[2,3],gender:2,gener:[2,3],generalizedentropyindex:2,get:[1,2],get_all_scor:[2,3,5],get_scor:[2,3,5],github:3,given:[1,2],grgic:[0,2],ground:2,group:[0,2],gummadi:[0,2],h:[0,1,2],ha:[1,2],hardt:0,hat:[0,1,2],have:1,head:[],heavili:3,heidari:[0,2],high:[1,2],higher:[1,2],highest:1,histor:[1,2],hlaca:[0,2],homogen:2,host:3,how:[1,2],howev:2,i:[0,1,2],ideal:[3,5],ideal_valu:[3,5],identifi:1,ignor:2,impact:3,implement:[0,1,2],improv:1,includ:2,incorrectli:2,index:[2,3],indic:[0,1,2],individu:[0,2],industri:1,inequ:[0,2],inform:[0,2],instal:3,instanc:[1,2],integr:2,intellig:3,inter:[2,3],interact:[1,2],interlistdivers:2,intern:[0,1,2],intra:2,intralistdivers:2,intuit:2,invers:[1,2],invest:3,ip:[1,2],irrelev:1,is_memb:[2,3,5],item:[1,2],item_featur:2,item_id:[2,3,5],item_id_column:2,itself:2,j:[1,2],john:[1,2],juriti:[1,4,5],k:[0,1,2,3,5],kept:2,knowledg:[0,2],known:2,label:[2,3,5],lack:[0,2],langford:[1,2],larg:1,last:1,latest:4,learn:[0,1,2,3,5],left:[0,1,2],less:1,let:[1,2],level:[1,2],li:[1,2],librari:0,lihong:[1,2],like:[1,2],likelihood:[2,3,5],line:1,linear:2,list:[2,3,4],list_of_class:2,ln:[0,2],log2:2,log:2,log_2:[1,2],lower:[3,5],lower_bound:[3,5],lowest:1,m:[0,1,2,4],m_i:1,machin:[1,2],made:[1,2],mai:2,make:2,map:[2,3,5],map_k:[3,5],match:[1,2],mathcal:[0,2],mathemat:0,mbox:[1,2],mean:[2,3],meanwhil:2,measur:[0,1,2],member:0,membership:2,membership_label:2,memori:2,men:2,method:[1,2],metric:0,mid:[0,2],min:[1,2],mine:[0,2],minibatch:2,minimum:2,miroslav:[1,2],mitig:0,mix:[3,5],model:1,modul:3,more:[1,2],moritz:0,most:1,mu:[0,2],multi:[2,3,5],multi_class_multi_label_predict:[3,5],multi_class_predict:[3,5],multi_metr:[3,5],multiclass:3,multiclassfairnessmetr:[2,3,5],multidisparateimpact:2,multipl:2,multistatisticalpar:2,must:2,n:[0,1,2,3,5],n_job:2,nan:2,nathan:0,ncdg:[3,5],ncdg_k:[3,5],ndarrai:2,ndcg:[2,3,5],ne:[0,2],necessari:4,need:2,neg:[0,2,3],neural:0,nip:0,non:[1,2],none:[2,4],normal:[2,3],notion:1,np:2,num_run:2,number:2,numpi:[2,4],object:2,observ:[1,2],odd:[2,3],offer:1,one:[1,2],onli:[1,2],oper:[1,2],opportun:[2,3],optim:0,option:2,other:[1,2,3,5],outcom:0,over:[1,2],overal:2,p:[0,1,2],p_:[1,2],p_i:[1,2],packag:4,pair:[1,2],panda:[2,3,4,5],paramet:2,pariti:3,part:2,pass:[2,4],pd:[2,3,5],pep:3,per:2,perform:[1,2],pinpoint:1,pip:3,platform:4,polici:[1,2],posit:[0,1,2],positive_label_nam:2,practic:[0,1,2],precis:[2,3,5],precision_i:[1,2],precision_k:[3,5],predict:[1,2,3,5],predicted_result:2,predictiveequ:2,preprint:[1,2],price:0,print:[2,3,5],probabl:[0,1,2],proceed:[0,1,2],process:[0,2],produc:1,propens:[1,2],propensity_column:2,propos:[0,2],protect:[0,2],provid:[1,2,3],put:1,py3:4,py:4,python:4,q:[1,2],qualiti:1,quantifi:[0,2],quick:4,r:[1,2],r_:[1,2],r_a:[1,2],rac:2,race:2,random:0,randomli:1,rang:[1,2],rank:3,rankingrecometr:[2,3,5],rate:[0,2,3,5],ratio:[0,2],recal:[2,3,5],recall_k:[3,5],receiv:[1,2],recommendations_batch:2,reduc:2,refer:[0,2],regul:2,rel:[1,2],relat:1,relev:[1,2],repo:4,report:2,repres:1,requir:3,research:3,respect:[0,1],result:[2,3,5],return_extended_result:2,reward:[1,2],right:[0,1,2],robust:[1,2],roc:2,row:2,run:[2,4],s:[1,2],same:[1,2],sampl:2,sample_weight:2,saw:[1,2],scenario:2,score:[1,2,3,5],scratch:4,second:[1,2],see:[0,2,3,5],seed:2,seen:2,self:2,sensit:1,seri:2,set:[1,2],setup:3,setuptool:4,should:[2,4],sigkdd:[0,2],sigma:2,singl:2,singla:[0,2],situat:[0,2],size:[1,2],some:[1,2],specif:2,specifi:2,speicher:[0,2],srebro:0,standard:[1,3,4],start:4,statist:3,statistical_par:2,statisticalpar:[2,3,5],str:2,subset:[0,1],success:4,sum:[1,2],sum_:[0,1,2],supervis:0,support:2,system:[0,1,3],t:[0,1,2],t_0:1,t_1:1,tabul:2,take:2,taken:1,task:0,techniqu:0,term:[1,2],test:[1,2,3,5],text:[0,2],th:[1,2],than:1,thei:1,theil:[2,3],theilindex:2,them:1,therefor:2,thereof:[0,2],thi:[0,1,2],three:[1,2],through:[2,3],time:[1,2],tnr:2,top:[1,2],total:[1,2],tp:[0,2],tpr:[0,2],tpr_:[0,2],train:[2,3,5],transform:[3,5],truth:2,tupl:2,two:[0,2,4],txt:4,type:2,u:[1,2],u_i:[1,2],u_j:[1,2],under:[2,3],unfair:[0,2],unifi:[0,2],uniform:0,union:2,uniqu:[1,2],unittest:4,upgrad:3,upper:[3,5],upper_bound:[3,5],us:[1,2,4],user:[1,2],user_id:[2,3,5],user_id_column:2,user_sample_s:2,v:4,v_p:[1,2],v_q:[1,2],valu:[0,1,2,3,5],value_column:2,varieti:2,variou:1,vector:1,version:[1,4],via:[0,2,3],wa:[1,2,4],want:2,we:2,weight:[1,2],weller:[0,2],were:[1,2],wheel:4,when:[0,1,2],where:[0,1,2],whether:[1,2],which:[1,2],whl:4,whole:[2,4],within:2,woman:2,working_memori:2,worst:2,would:1,x:[0,1,2,4],y:[0,2],y_i:[0,2],you:[2,4],your:3,zafar:[0,2]},titles:["About Algorithmic Fairness","About Recommenders Metrics","Jurity Public API","Jurity: Fairness &amp; Evaluation Library","Installation","Quick Start"],titleterms:{"public":2,about:[0,1],algorithm:0,api:2,appli:[3,5],area:1,auc:1,avail:3,averag:[0,1],bia:[2,3,5],binari:[1,2],calcul:[3,5],classif:[2,3,5],click:1,code:[3,4],content:3,ctr:1,cumul:1,curv:1,differ:0,discount:1,dispar:0,divers:[1,2],entropi:0,equal:0,evalu:3,fair:[0,2,3,5],fit:[3,5],fnr:0,from:4,gain:1,gener:0,impact:0,index:0,indic:3,instal:4,inter:1,intra:1,juriti:[2,3],librari:[3,4],list:1,map:1,mean:1,metric:[1,2,3,5],mitig:[2,3,5],multiclass:2,ndcg:1,normal:1,odd:0,opportun:0,option:4,pariti:0,pip:4,precis:1,predict:0,quick:[3,5],rank:[1,2],rate:1,recal:1,recommend:[1,2,3,5],requir:4,setup:4,sourc:[3,4],start:[3,5],statist:0,tabl:3,techniqu:3,test:4,theil:0,through:1,under:1,upgrad:4,via:4,your:4}})
+Search.setIndex({docnames:["about_fairness","about_reco","api","index","install","quick"],envversion:{"sphinx.domains.c":2,"sphinx.domains.changeset":1,"sphinx.domains.citation":1,"sphinx.domains.cpp":4,"sphinx.domains.index":1,"sphinx.domains.javascript":2,"sphinx.domains.math":2,"sphinx.domains.python":3,"sphinx.domains.rst":2,"sphinx.domains.std":2,sphinx:56},filenames:["about_fairness.rst","about_reco.rst","api.rst","index.rst","install.rst","quick.rst"],objects:{"":[[2,0,0,"-","jurity"]],"jurity.classification":[[2,1,1,"","BinaryClassificationMetrics"]],"jurity.classification.BinaryClassificationMetrics":[[2,1,1,"","AUC"],[2,1,1,"","Accuracy"],[2,1,1,"","F1"],[2,1,1,"","Precision"],[2,1,1,"","Recall"]],"jurity.classification.BinaryClassificationMetrics.AUC":[[2,2,1,"","get_score"]],"jurity.classification.BinaryClassificationMetrics.Accuracy":[[2,2,1,"","get_score"]],"jurity.classification.BinaryClassificationMetrics.F1":[[2,2,1,"","get_score"]],"jurity.classification.BinaryClassificationMetrics.Precision":[[2,2,1,"","get_score"]],"jurity.classification.BinaryClassificationMetrics.Recall":[[2,2,1,"","get_score"]],"jurity.fairness":[[2,1,1,"","BinaryFairnessMetrics"],[2,1,1,"","MultiClassFairnessMetrics"]],"jurity.fairness.BinaryFairnessMetrics":[[2,1,1,"","AverageOdds"],[2,3,1,"","DisparateImpact"],[2,1,1,"","EqualOpportunity"],[2,1,1,"","FNRDifference"],[2,1,1,"","GeneralizedEntropyIndex"],[2,1,1,"","PredictiveEquality"],[2,3,1,"","StatisticalParity"],[2,1,1,"","TheilIndex"],[2,2,1,"","get_all_scores"]],"jurity.fairness.BinaryFairnessMetrics.AverageOdds":[[2,2,1,"","get_score"]],"jurity.fairness.BinaryFairnessMetrics.EqualOpportunity":[[2,2,1,"","get_score"]],"jurity.fairness.BinaryFairnessMetrics.FNRDifference":[[2,2,1,"","get_score"]],"jurity.fairness.BinaryFairnessMetrics.GeneralizedEntropyIndex":[[2,2,1,"","get_score"]],"jurity.fairness.BinaryFairnessMetrics.PredictiveEquality":[[2,2,1,"","get_score"]],"jurity.fairness.BinaryFairnessMetrics.TheilIndex":[[2,2,1,"","get_score"]],"jurity.fairness.MultiClassFairnessMetrics":[[2,3,1,"","DisparateImpact"],[2,3,1,"","StatisticalParity"],[2,2,1,"","get_all_scores"]],"jurity.mitigation":[[2,1,1,"","BinaryMitigation"]],"jurity.mitigation.BinaryMitigation":[[2,1,1,"","EqualizedOdds"]],"jurity.mitigation.BinaryMitigation.EqualizedOdds":[[2,2,1,"","fit"],[2,2,1,"","fit_transform"],[2,2,1,"","transform"]],"jurity.recommenders":[[2,1,1,"","BinaryRecoMetrics"],[2,1,1,"","DiversityRecoMetrics"],[2,1,1,"","RankingRecoMetrics"]],"jurity.recommenders.BinaryRecoMetrics":[[2,1,1,"","AUC"],[2,1,1,"","CTR"]],"jurity.recommenders.BinaryRecoMetrics.AUC":[[2,2,1,"","get_score"]],"jurity.recommenders.BinaryRecoMetrics.CTR":[[2,2,1,"","get_score"]],"jurity.recommenders.DiversityRecoMetrics":[[2,1,1,"","InterListDiversity"]],"jurity.recommenders.DiversityRecoMetrics.InterListDiversity":[[2,2,1,"","get_score"]],"jurity.recommenders.RankingRecoMetrics":[[2,1,1,"","MAP"],[2,1,1,"","NDCG"],[2,1,1,"","Precision"],[2,1,1,"","Recall"]],"jurity.recommenders.RankingRecoMetrics.MAP":[[2,2,1,"","get_score"]],"jurity.recommenders.RankingRecoMetrics.NDCG":[[2,2,1,"","get_score"]],"jurity.recommenders.RankingRecoMetrics.Precision":[[2,2,1,"","get_score"]],"jurity.recommenders.RankingRecoMetrics.Recall":[[2,2,1,"","get_score"]]},objnames:{"0":["py","module","Python module"],"1":["py","class","Python class"],"2":["py","method","Python method"],"3":["py","attribute","Python attribute"]},objtypes:{"0":"py:module","1":"py:class","2":"py:method","3":"py:attribute"},terms:{"0":[0,1,2,3,5],"1":[0,1,2,3,5],"10":2,"1000":[],"10000":2,"100000":[],"1103":[1,2],"12":2,"122":2,"16":0,"2":[0,2,3,5],"2011":[1,2],"2016":[0,2],"2018":[0,2],"28th":[1,2],"29":2,"3":[2,3,4,5],"30":[3,5],"30th":0,"316":2,"4":[2,3,5],"453":2,"4601":[1,2],"5":[1,3,5],"59495d2077d4":[],"6":[3,4,5],"65":2,"68":2,"7":[3,5],"8":[3,5],"boolean":[1,2],"case":[0,1,2],"class":[2,3,5],"default":2,"dud\u00edk":[1,2],"float":2,"function":[1,2],"import":[1,2,3,5],"int":2,"new":2,"public":3,"return":[1,2],"static":2,"true":[0,2],"while":[1,2],A:[0,1,2],At:[1,2],For:1,If:[1,2],In:[0,1,2],It:[1,2],That:2,The:[1,2,3,4],There:[1,2,4],To:4,With:[0,2],__k__:1,_a:1,_basebinaryfair:2,_basemitig:2,_baserecommend:2,_distanc:[1,2],_i:[0,2],_item_id_column:2,_similar:[],_user_id_column:2,a_i:[1,2],abil:2,abl:1,about:3,abov:2,account:[1,2],accumul:2,accuraci:[0,1,2,3,5],achiev:[0,2],acm:[0,2],across:[0,2],actual:[1,2,3,5],actual_respons:2,actual_responses_batch:2,actual_result:2,addit:2,adher:3,advanc:2,after:[3,4,5],agre:1,algorithm:[2,3],alia:2,all:[1,2,4],allow:2,alpha:[0,2],also:[1,2,3,5],altern:4,alwai:1,among:2,amount:1,an:[1,2],analog:2,ani:[2,4],anyth:2,api:3,appear:1,appli:2,approach:[0,2],ar:[1,2,4],area:[2,3],artifici:3,arxiv:[1,2],associ:2,assum:2,attribut:[0,2],auc:[2,3,5],auc_acc:2,auc_batch:2,auxiliari:2,avail:[1,2,5],averag:[2,3],averageodd:2,avoid:1,b:[0,2,3,5],b_:[0,2],b_i:[0,2],balanc:2,base:[1,2],batch:2,batch_accumul:2,bdist_wheel:4,befor:[3,5],begin:[0,2],being:[1,2],below:[0,1],best:2,between:[0,1,2],bia:0,binari:[0,3,5],binary_predict:[3,5],binaryclassificationmetr:[2,3,5],binarydisparateimpact:2,binaryfairnessmetr:[2,3,5],binarymitig:[2,3,5],binaryrecometr:[2,3,5],binarystatisticalpar:2,black:2,bool:2,both:1,bound:[3,5],box:2,build:4,c:[1,3,5],calcul:[0,1,2],call:2,callabl:2,can:[1,2,4],cannot:2,cap:[1,2],captur:[1,2],cdot:[1,2],center:3,characterist:2,chosen:1,chunk_siz:[],classif:0,classifi:[0,2],click:[2,3,5],click_column:[2,3,5],clone:4,close:2,column:2,com:[],combin:[1,2],come:1,command:4,commun:1,comput:2,condit:[0,2],confer:[0,1,2],confirm:4,consist:[1,2],constrain:0,contain:2,context:[1,2],contribut:2,control:2,convex:2,core:2,correct:[1,2],correctli:2,cosin:[1,2],cosine_dist:[],cosine_similar:[],ctr:[2,3,5],ctr_acc:2,ctr_batch:2,cumul:[2,3],current:2,curv:[2,3],d:[0,2],data:[0,1,2,3,5],datafram:[2,3,5],dataset:[0,2],decis:[0,2],defin:[0,2],definit:[0,1],denot:[0,1,2],depend:[2,4],descript:[3,5],design:2,determin:2,develop:3,df56c6611093:[],dict:2,didn:1,differ:[1,2,3],direct:[1,2],directli:[1,2],discount:[2,3],discov:4,discoveri:[0,2],dispar:3,disparate_impact:2,disparateimpact:2,dist:4,distanc:2,distribut:2,divers:3,diversifi:[1,2],diversityrecometr:2,divid:1,done:1,doubli:[1,2],dr:[1,2],draw:2,due:1,e:[0,2],each:[0,1,2],easili:2,edg:1,effect:1,either:2,els:2,end:[0,2],enough:1,entropi:[2,3],equal:[2,3],equalizedodd:[2,3,5],equalopportun:2,equat:2,eric:[0,2],error:1,especi:1,estim:[1,2],evalu:[1,2],even:1,exampl:[0,2,4],excel:3,ext:[],extend:2,extrm:[],f1:[2,3,5],f1_score:[3,5],f:[1,2],fair_likelihood:[2,3,5],fair_predict:[2,3,5],fairer:2,fals:[0,2,3],feedback:[],fidel:3,filter:1,find:2,first:4,fit:2,fit_transform:2,flip:[0,2],fn:2,fnr:[2,3],fnrdiffer:2,follow:4,footprint:2,foral:[1,2],formal:2,formula:2,four:2,fp:2,fpr:[0,2],fpr_:[0,2],frac:[0,1,2],fraction:2,frame:2,from:[1,2,3,5],full:2,further:1,g:[0,2],gain:[2,3],gender:2,gener:[2,3],generalizedentropyindex:2,get:[1,2],get_all_scor:[2,3,5],get_scor:[2,3,5],github:3,given:[1,2],grgic:[0,2],ground:2,group:[0,2],gummadi:[0,2],h:[0,1,2],ha:[1,2],hardt:[0,2],hat:[0,1,2],have:[1,2],heavili:3,heidari:[0,2],here:2,high:[1,2],higher:[1,2],highest:1,histor:[1,2],hlaca:[0,2],homogen:2,host:3,how:[1,2],howev:2,http:[],hull:2,i:[0,1,2],idea:2,ideal:[3,5],ideal_valu:[3,5],identifi:1,ignor:2,im:[],imagin:2,impact:3,implement:[0,1,2],implicit:[],improv:1,includ:2,incorrectli:2,index:[2,3],indic:[0,1,2],individu:[0,2],industri:1,inequ:[0,2],inform:[0,2],instal:3,instanc:[1,2],integr:2,intellig:3,inter:[2,3],interact:[1,2],interlist:[],interlistdivers:2,intern:[0,1,2],intuit:2,invers:[1,2],invest:3,ip:[1,2],irrelev:1,is_memb:[2,3,5],item:[1,2],item_id:[2,3,5],item_id_column:2,itself:2,j:[1,2],john:[1,2],judaikawa:[],juriti:[1,4,5],k:[0,1,2,3,5],kept:2,knowledg:[0,2],known:2,label:[2,3,5],lack:[0,2],langford:[1,2],larg:1,last:1,latest:4,learn:[0,1,2,3,5],left:[0,1,2],less:1,let:[1,2],level:[1,2],li:[1,2],librari:0,lihong:[1,2],like:[1,2],likelihood:[2,3,5],line:1,linear:2,list:[2,3,4],list_of_class:2,ln:[0,2],log2:2,log:2,log_2:[1,2],lower:[2,3,5],lower_bound:[3,5],lowest:1,m:[0,1,2,4],m_i:1,machin:[1,2],made:[1,2],mai:2,make:2,map:[2,3,5],map_k:[3,5],match:[1,2],math:[],mathcal:[0,2],mathemat:0,mbox:[1,2],mean:[2,3],measur:[0,1,2],medium:[],member:0,membership:2,membership_label:2,memori:2,men:2,method:[1,2],metric:0,mid:[0,2],might:2,min:[1,2],mine:[0,2],minibatch:2,minimum:2,miroslav:[1,2],mitig:0,mix:[2,3,5],model:1,modul:3,more:[1,2],moritz:[0,2],most:1,mu:[0,2],multi:[2,3,5],multi_class_multi_label_predict:[3,5],multi_class_predict:[3,5],multi_metr:[3,5],multiclass:3,multiclassfairnessmetr:[2,3,5],multidisparateimpact:2,multipl:2,multistatisticalpar:2,must:2,n:[0,1,2,3,5],n_job:2,nan:2,nathan:0,nati:2,ncdg:[3,5],ncdg_k:[3,5],ndarrai:2,ndcg:[2,3,5],ne:[0,2],necessari:4,need:2,neg:[0,2,3],neural:[0,2],nip:0,non:[1,2],none:[2,4],normal:[2,3],notion:1,np:2,num_run:2,number:2,numpi:[2,4],object:2,observ:[1,2],odd:[2,3],off:2,offer:1,one:[1,2],onli:[1,2],oper:[1,2],opportun:[2,3],optim:[0,2],option:2,oral:[],other:[1,2,3,5],outcom:[0,2],output:2,over:1,overal:2,p:[0,1,2],p_:[1,2],p_i:[1,2],packag:4,pair:[1,2],panda:[2,3,4,5],paramet:2,pariti:3,part:2,pass:[2,4],pd:[2,3,5],pep:3,per:2,perform:[1,2],pinpoint:1,pip:3,platform:4,polici:[1,2],posit:[0,1,2],positive_label_nam:2,possibl:2,practic:[0,1,2],precis:[2,3,5],precision_i:[1,2],precision_k:[3,5],predict:[1,2,3,5],predicted_result:2,predictiveequ:2,predictor:2,preprint:[1,2],price:[0,2],print:[2,3,5],probabilist:2,probabl:[0,1,2],proceed:[0,1,2],process:[0,2],produc:1,propens:[1,2],propensity_column:2,propos:[0,2],protect:[0,2],provid:[1,2,3],put:1,py3:4,py:4,python:4,qualiti:1,quantifi:[0,2],quick:4,r:[1,2],r_:[1,2],r_a:[1,2],rac:[],race:2,random:[0,2],randomli:1,rang:[1,2],rank:3,rankingrecometr:[2,3,5],rate:[0,2,3,5],ratio:[0,2],recal:[2,3,5],recall_k:[3,5],receiv:[1,2],recommendations_batch:2,reduc:2,refer:[0,2],regul:2,rel:[1,2],relat:1,relationship:2,relev:[1,2],repair:2,repo:4,repres:[1,2],requir:3,research:3,respect:[0,1],result:[2,3,5],return_extended_result:2,reward:[1,2],right:[0,1,2],robust:[1,2],roc:2,row:2,run:4,s:[1,2],same:[1,2],sampl:2,sample_s:[],sample_weight:2,satisfi:2,saw:[1,2],scenario:2,score:[1,2,3,5],scratch:4,second:[1,2],see:[0,2,3,5],seed:2,seen:2,self:2,sensit:1,seri:2,set:[1,2],setup:3,setuptool:4,should:[2,4],sigkdd:[0,2],sigma:2,singl:2,singla:[0,2],situat:[0,2],size:[1,2],so:2,some:[1,2],sourc:[],specif:2,specifi:2,speicher:[0,2],srebro:[0,2],standard:[1,3,4],start:4,statist:3,statistical_par:2,statisticalpar:[2,3,5],str:2,subset:[0,1,2],success:4,sum:[1,2],sum_:[0,1,2],supervis:[0,2],support:2,system:[0,1,2,3],t:[0,1,2],t_0:1,t_1:1,tabul:2,take:2,taken:1,task:[0,2],techniqu:0,term:[1,2],test:[1,2,3,5],text:[0,2],th:[1,2],than:1,thei:1,theil:[2,3],theilindex:2,them:1,therefor:2,thereof:[0,2],thi:[0,1,2],three:[1,2],through:[2,3],time:[1,2],tnr:2,top:[1,2],total:[1,2],towardsdatasci:[],tp:[0,2],tpr:[0,2],tpr_:[0,2],trade:2,train:[2,3,5],transform:[2,3,5],truth:2,tupl:2,two:[0,2,4],txt:4,type:2,u:[1,2],u_i:[1,2],u_j:[1,2],under:[2,3],unfair:[0,2],unifi:[0,2],uniform:0,union:2,uniqu:[1,2],unittest:4,unseen:2,upgrad:3,upper:[3,5],upper_bound:[3,5],us:[1,2,4],user:[1,2],user_id:[2,3,5],user_id_column:2,user_sample_s:2,v:4,valu:[0,1,2,3,5],value_column:2,varieti:2,variou:1,vector:[1,2],version:[1,4],via:[0,2,3],wa:[1,2,4],want:2,we:2,weight:[1,2],weller:[0,2],were:[1,2],wheel:4,when:[0,1,2],where:[0,1,2],whether:[1,2],which:[1,2],whl:4,whole:[2,4],within:2,without:2,woman:2,working_memori:2,worst:2,would:1,x:[0,1,2,4],y:[0,2],y_i:[0,2],you:[2,4],your:3,zafar:[0,2]},titles:["About Algorithmic Fairness","About Recommenders Metrics","Jurity Public API","Jurity: Fairness &amp; Evaluation Library","Installation","Quick Start"],titleterms:{"public":2,about:[0,1],algorithm:0,api:2,appli:[3,5],area:1,auc:1,avail:3,averag:[0,1],bia:[2,3,5],binari:[1,2],calcul:[3,5],classif:[2,3,5],click:1,code:[3,4],content:3,ctr:1,cumul:1,curv:1,differ:0,discount:1,dispar:0,divers:[1,2],entropi:0,equal:0,evalu:3,fair:[0,2,3,5],fit:[3,5],fnr:0,from:4,gain:1,gener:0,impact:0,index:0,indic:3,instal:4,inter:1,juriti:[2,3],librari:[3,4],list:1,map:1,mean:1,metric:[1,2,3,5],mitig:[2,3,5],multiclass:2,ndcg:1,normal:1,odd:0,opportun:0,option:4,pariti:0,pip:4,precis:1,predict:0,quick:[3,5],rank:[1,2],rate:1,recal:1,recommend:[1,2,3,5],requir:4,setup:4,sourc:[3,4],start:[3,5],statist:0,tabl:3,techniqu:3,test:4,theil:0,through:1,under:1,upgrad:4,via:4,your:4}})
