+++ jurity/recommenders/__init__.py
@@ -7,7 +7,7 @@ from .map import MAP
 from .ndcg import NDCG
 from .precision import Precision
 from .recall import Recall
-from .diversity import InterListDiversity, IntraListDiversity
+from .interlist_diversity import InterListDiversity
 
 
 class BinaryRecoMetrics(NamedTuple):
@@ -24,4 +24,3 @@ class RankingRecoMetrics(NamedTuple):
 
 class DiversityRecoMetrics(NamedTuple):
     InterListDiversity = InterListDiversity
-    IntraListDiversity = IntraListDiversity
